requirements.txt
gitdb==4.0.9
GitPython==3.1.26
smmap==5.0.0